+++ TicketManager/TicketManager.DataAccess.Events/DataModel/TicketStatusChangedEvent.cs
@@ -5,7 +5,7 @@ namespace TicketManager.DataAccess.Events.DataModel
 {
     public class TicketStatusChangedEvent : EventBase, ITicketEvent
     {
-        public int TicketCreatedEventId { get; set; }
+        public long TicketCreatedEventId { get; set; }
 
         [ForeignKey(nameof(TicketCreatedEventId))]
         public virtual TicketCreatedEvent TicketCreatedEvent { get; set; }
