+++ TicketManager/TicketManager.WebAPI/DTOs/Commands/Abstractions/TicketCommandBase.cs
@@ -4,7 +4,7 @@ namespace TicketManager.WebAPI.DTOs.Commands.Abstractions
 {
     public abstract class TicketCommandBase : TicketCommandBase<Unit>
     {
-        protected TicketCommandBase(int ticketId, string raisedByUser)
+        protected TicketCommandBase(long ticketId, string raisedByUser)
             : base(ticketId, raisedByUser)
         {
         }
